Sructure and readability improvements

+++ handlers/middlewares/checkLinks.js
@@ -17,37 +17,32 @@ const reply_markup = { inline_keyboard: warnInlineKeyboard };
 
 const bannedDomains = new Set([
 	'chat.whatsapp.com',
+	'loverdesuootnosheniya.blogspot.com',
 	't.me',
 	'telegram.dog',
 	'telegram.me',
-	'tinyurl.com',
+	'your-sweet-dating.com',
 ]);
 
 
-const ClassifiedSync = taggedSum('Listed', {
-	Neither: [ 'url' ],
-	Warn: [ 'blacklist' ],
-	// Support whitelist?
-});
-
-const ReplyType = taggedSum('ReplyType', {
-	Error: [ 'errorMsg' ],
-	Ok: [ 'url' ],
-});
-
 const Action = taggedSum('Action', {
 	Nothing: [],
-	NotifyFetchError: [ 'errorMsg' ],
+	Notify: [ 'errorMsg' ],
 	Warn: [ 'reason' ],
 });
 
+const cata = R.invoker(1, 'cata');
+const actionPriority = cata({
+	// Numbers are relative to each other
+	Nothing: () => 0,
+	Notify: () => 10,
+	Warn: () => 20,
+});
+const maxByActionPriority = R.maxBy(actionPriority);
+const highestPriorityAction = R.reduce(maxByActionPriority, Action.Nothing);
 
 const assumeProtocol = R.unless(R.contains('://'), R.concat('http://'));
-const cata = R.invoker(1, 'cata');
 const constructAbsUrl = R.constructN(1, URL);
-const getTag = R.prop('@@tag');
-const getUrl = R.prop('url');
-const isClassifiedWarn = syncCls => ClassifiedSync.Warn.is(syncCls);
 const isHttp = R.where({ protocol: R.test(/https?:/) });
 const isUrl = entity => entity.type === 'url' || entity.type === 'text_link ';
 const memoize = R.memoizeWith(R.identity);
@@ -58,47 +53,30 @@ const obtainUrlFromText = text => ({ length, offset, url }) =>
 		: text.slice(offset, length + offset);
 
 
-const classifySync1 = url => {
-	// Support whitelist?
-	if (url.pathname !== '/' && bannedDomains.has(url.host.toLowerCase())) {
-		return ClassifiedSync.Warn('domain');
-	}
-
-	if (url.protocol === 'tg:' && url.host.toLowerCase() === 'resolve') {
-		return ClassifiedSync.Warn('protocol');
-	}
-
-	return ClassifiedSync.Neither(url);
+const blacklisted = {
+	domain: url =>
+		bannedDomains.has(url.host.toLowerCase()),
+	protocol: url =>
+		url.protocol === 'tg:' && url.host.toLowerCase() === 'resolve',
 };
 
-const classifyAsync = memoize(url =>
-	fetch(url, { redirect: 'follow' }).then(res => {
-		if (res.ok) return ReplyType.Ok(new URL(res.url));
+const unshorten = memoize(url =>
+	fetch(url, { redirect: 'follow' }).then(res =>
+		res.ok
+			? new URL(res.url)
+			: Promise.reject(new Error(`Request to ${url} failed, ` +
+				`reason: ${res.status} ${res.statusText}`))));
 
-		return ReplyType.Error(`Request to ${url} failed, ` +
-			`reason: ${res.status} ${res.statusText}`);
-	}).catch(ReplyType.Error));
+const classifyAsync = url =>
+	unshorten(url).then(long => blacklisted.domain(long)
+		? Action.Warn('blacklisted domain')
+		: Action.Nothing).catch(Action.Notify);
 
-// Feel free to improve readability of this functions, it'll benefit from it.
-// Naming in the whole file also isn't the best.
-// ...I hope it's all readable enough.
-const classifyList = async (urls) => {
-	const { Warn, Neither = [] } = R.groupBy(getTag, urls.map(classifySync1));
+const classifyList = (urls) => {
+	if (urls.some(blacklisted.protocol)) return Action.Warn('tg: protocol');
 
-	if (Warn) return Action.Warn(Warn[0].blacklist);
-
-	// eslint-disable-next-line max-len
-	const resolved = await Promise.all(Neither.map(getUrl).filter(isHttp).map(classifyAsync));
-
-	const { Ok = [], Error } = R.groupBy(getTag, resolved);
-
-	const warnCls = Ok.map(getUrl).map(classifySync1).find(isClassifiedWarn);
-
-	if (warnCls) return Action.Warn(warnCls.blacklist);
-
-	if (Error) return Action.NotifyFetchError(Error[0].errorMsg);
-
-	return Action.Nothing;
+	return Promise.all(urls.filter(isHttp).map(classifyAsync))
+		.then(highestPriorityAction);
 };
 
 
@@ -129,7 +107,7 @@ module.exports = async (ctx, next) => {
 	const urls = R.uniq(rawUrls).map(constructAbsUrl);
 
 	// if one link is repeated 3 times or more
-	if (rawUrls.length - urls.length >= 2) {
+	if (rawUrls.length - urls.length >= 2 && !await isAdmin(userToWarn)) {
 		const reason = 'Link - multiple copies, possibly spam';
 		ctx.deleteMessage();
 		const warnMessage = await warn({ admin, reason, userToWarn });
@@ -138,15 +116,15 @@ module.exports = async (ctx, next) => {
 
 	// TODO Whitelist invite links somewhere around here
 	// Could support whitelist here as well
-	return classifyList(urls).then(cata({
+	return (await classifyList(urls)).cata({
 		Nothing: next,
-		NotifyFetchError(errorMsg) {
+		Notify(errorMsg) {
 			const reply_to_message_id = message.message_id;
 			ctx.reply(`️ℹ️ ${errorMsg}`, { reply_to_message_id });
 			return next();
 		},
 		Warn: async (blacklist) => {
-			const reason = `Link - blacklisted ${blacklist}`;
+			const reason = `Link - ${blacklist}`;
 
 			if (await isAdmin(userToWarn)) return next();
 
@@ -154,5 +132,5 @@ module.exports = async (ctx, next) => {
 			const warnMessage = await warn({ admin, reason, userToWarn });
 			return ctx.replyWithHTML(warnMessage, { reply_markup });
 		},
-	}));
+	});
 };
